requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
strawberry-graphql==0.284.1
requests==2.32.5
Flask==3.1.2
grpcio==1.76.0
grpcio-tools==1.76.0
aio-pika==9.5.8
PyJWT==2.10.1
asyncio==3.4.3
confluent-kafka==2.12.2